Assingment1
The objective of this assignment is to learn about simple data curation
practices, and familiarize you with some of the data we'll be reusing
later.
This notebook uses the notMNIST dataset to be used with python
experiments. This dataset is designed to look like the classic MNIST
dataset, while looking a little more like real data: it's a harder
task, and the data is a lot less 'clean' than MNIST.